# src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8090
  description: Generated server url
...
paths:
...
  /api/directors:
    get:
      tags:
      - director-entity-controller
      description: get-director
      operationId: getCollectionResource-director-get_1
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          default: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          minimum: 1
          type: integer
          default: 20
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PagedModelEntityModelDirector'
            application/x-spring-data-compact+json:
              schema:
                $ref: '#/components/schemas/PagedModelEntityModelDirector'
            text/uri-list:
              schema:
                type: string
    post:
      tags:
      - director-entity-controller
      description: create-director
      operationId: postCollectionResource-director-post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectorRequestBody'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDirector'
...
  /api/directors/{id}:
    get:
      tags:
      - director-entity-controller
      description: get-director
      operationId: getItemResource-director-get
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDirector'
        "404":
          description: Not Found
    put:
      tags:
      - director-entity-controller
      description: update-director
      operationId: putItemResource-director-put
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectorRequestBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDirector'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDirector'
        "204":
          description: No Content
    delete:
      tags:
      - director-entity-controller
      description: delete-director
      operationId: deleteItemResource-director-delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - director-entity-controller
      description: patch-director
      operationId: patchItemResource-director-patch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectorRequestBody'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelDirector'
        "204":
          description: No Content
...
  /api/titles/{id}/director:
    get:
      tags:
      - title-property-reference-controller
      description: get-director-by-title-Id
      operationId: followPropertyReference-title-get_3_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
            text/uri-list:
              schema:
                type: string
        "404":
          description: Not Found
    put:
      tags:
      - title-property-reference-controller
      description: update-director-by-title-Id
      operationId: createPropertyReference-title-put_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
        "201":
          description: Created
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
        "204":
          description: No Content
    delete:
      tags:
      - title-property-reference-controller
      description: delete-director-by-title-Id
      operationId: deletePropertyReference-title-delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
    patch:
      tags:
      - title-property-reference-controller
      description: patch-director-by-title-Id
      operationId: createPropertyReference-title-patch_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          application/x-spring-data-compact+json:
            schema:
              $ref: '#/components/schemas/CollectionModelObject'
          text/uri-list:
            schema:
              type: string
        required: true
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
        "204":
          description: No Content
  /api/titles/{id}/director/{propertyId}:
    get:
      tags:
      - title-property-reference-controller
      description: get-director-by-title-Id
      operationId: followPropertyReference-title-get_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelDirector'
        "404":
          description: Not Found
    delete:
      tags:
      - title-property-reference-controller
      description: delete-director-by-title-Id
      operationId: deletePropertyReferenceId-title-delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
...
  /api/recommend/category/{id}:
    get:
      tags:
      - recommend-controller
      operationId: bestInCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: top
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
  /api/recommend/best:
    get:
      tags:
      - recommend-controller
      operationId: bestTitles
      parameters:
      - name: top
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            application/hal+json:
              schema:
                type: object
components:
  schemas:
...
    Link:
      type: object
      properties:
        href:
          type: string
        hreflang:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
        templated:
          type: boolean
...
    PageMetadata:
      type: object
      properties:
        size:
          type: integer
          format: int64
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int64
        number:
          type: integer
          format: int64
...
    CollectionModelDirector:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            directors:
              type: array
              items:
                $ref: '#/components/schemas/DirectorResponse'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    CollectionModelObject:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            objects:
              type: array
              items:
                type: object
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
...
    EntityModelDirector:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    PagedModelEntityModelDirector:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            directors:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelDirector'
        _links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        page:
          $ref: '#/components/schemas/PageMetadata'
...
    DirectorRequestBody:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        title:
          type: array
          items:
            type: string
...
    DirectorResponse:
      required:
      - name
      type: object
      properties:
        name:
          type: string
...
  securitySchemes:
    netflixAuth:
      type: http
      in: header
      scheme: basic